data 모듈 세팅
