--- 2003/tjdwlsl1021/app/src/main/java/com/example/myapplication/data/remote/source/MovieRemoteDataSource.kt
@@ -1,11 +1,0 @@
-package com.example.myapplication.data.remote.source
-
-import com.example.myapplication.data.local.MovieEntity
-
-interface MovieRemoteDataSource {
-    fun getMovieList(
-        query: String,
-        success: (List<MovieEntity>) -> Unit,
-        failed: (Throwable) -> Unit
-    )
-}
