+++ 2003/tjdwlsl1021/data2/src/main/java/com/example/data2/MovieRepository.kt
@@ -1,0 +1,11 @@
+package com.example.data2
+
+import com.example.data2.model.Movie
+
+interface MovieRepository {
+    fun getMovieList(
+        query: String,
+        success: (List<Movie>) -> Unit,
+        failed: (Throwable) -> Unit
+    )
+}
