+++ 2003/tjdwlsl1021/data2/src/main/java/com/example/data2/di/RepositoryModule.kt
@@ -1,0 +1,9 @@
+package com.example.data2.di
+
+import com.example.data2.MovieRepository
+import com.example.data2.MovieRepositoryImpl
+import org.koin.dsl.module
+
+val repositoryModule = module {
+    single<MovieRepository> { MovieRepositoryImpl(get(), get()) }
+}
