+++ 2003/tjdwlsl1021/data2/src/main/java/com/example/data2/model/Movie.kt
@@ -1,0 +1,12 @@
+package com.example.data2.model
+
+data class Movie(
+    val actor: String,
+    val director: String,
+    val image: String,
+    val link: String,
+    val pubDate: String,
+    val subtitle: String,
+    val title: String,
+    val userRating: Float
+)
