+++ 2003/tjdwlsl1021/data2/src/main/java/com/example/data2/soruce/local/MovieLocalDataSource.kt
@@ -1,0 +1,8 @@
+package com.example.data2.soruce.local
+
+import com.example.data2.model.Movie
+
+interface MovieLocalDataSource {
+    fun insertMovies(movies: List<Movie>)
+    fun getSearchMovies(title: String): List<Movie>
+}
